// frontend/apps/talent/apis/model/profileUpdateResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Talent Service
 * talent APIs
 * OpenAPI spec version: 1.0
 */
import type { Profile } from "./profile";
import type { ProfileUpdateValidationError } from "./profileUpdateValidationError";

export interface ProfileUpdateResponse {
  profile: Profile;
  errors: ProfileUpdateValidationError;
}