// frontend/apps/talent/apis/model/profileUpdateValidationError.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Talent Service
 * talent APIs
 * OpenAPI spec version: 1.0
 */

export interface ProfileUpdateValidationError {
  firstName?: string[];
  lastName?: string[];
  birthday?: string[];
  address?: string[];
  currentEmployer?: string[];
  tel?: string[];
}